Merges all hooks into a single hook file. (#37)
Description of changes:
Merges all hooks in each resource into a single hook file.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.
